Include shokkenki provider in RSpec backtrace.
Because the generated examples have no file and
line number, nothing is added to the backtrace
when an example fails.

RSpec's default backtrace cleaning then cleans
the entire backtrace and in turn decides to show
the entire backtrace, resulting in a lot of crap
when an example fails.

Including the shokkenki-provider lines ensures
that at least some backtrace is present after the
clean which then allows a smaller backtrace to be
displayed. Confused? Yes.

A better solution would be to add a line to the
backtrace for our failing examples but I have no
idea of how hard that is or whether it is worth
digging around in the RSpec internals for.